delete old MATCH controllers
